Attempt to config_from_object twice
in theory -- the inital config_ will occur on any import -- this will allow for other scripts to interact with the appropriate instance. However, I'm not sure this will work since the queues are not set up. If this doesn't work, I'll be attempting to import the app from a function