Average data, use power pin
The device runs the fan nonstop, will use a pin to cut power to device when not in use. Will also take 3 measurements and average the results. NOTE: the sensor requires a 30sec initialization and the sample freq is 2.3 seconds and so there is a delay (5 sec in this case, for two cycles) before readings